refactor: move web3 outside of "shared" (#4566)
