+++ browser-interface/packages/lib/web3/catalystABI.ts
@@ -58,5 +58,5 @@ export const catalystABI = [
     payable: false,
     stateMutability: 'view',
     type: 'function'
-  },
-];
+  }
+]
