+++ browser-interface/packages/shared/friends/sagas.ts
@@ -49,6 +49,7 @@ import { deepEqual } from 'lib/javascript/deepEqual'
 import { now } from 'lib/javascript/now'
 import { uuid } from 'lib/javascript/uuid'
 import defaultLogger, { createDummyLogger, createLogger } from 'lib/logger'
+import { fetchENSOwner } from 'lib/web3/fetchENSOwner'
 import { apply, call, delay, fork, put, race, select, take, takeEvery } from 'redux-saga/effects'
 import { trackEvent } from 'shared/analytics/trackEvent'
 import { SendPrivateMessage, SEND_PRIVATE_MESSAGE } from 'shared/chat/actions'
@@ -153,7 +154,6 @@ import {
   UpdateTotalUnseenMessagesPayload,
   UpdateUserUnseenMessagesPayload
 } from 'shared/types'
-import { fetchENSOwner } from 'shared/web3'
 import { getUnityInstance } from 'unity-interface/IUnityInterface'
 import { ensureFriendProfile } from './ensureFriendProfile'
 import {
