+++ browser-interface/packages/shared/profiles/sagas/initialRemoteProfileLoad.ts
@@ -5,7 +5,7 @@ import { call, put, select } from 'redux-saga/effects'
 import { BringDownClientAndReportFatalError, ErrorContext } from 'shared/loading/ReportFatalError'
 import { getCurrentIdentity, getCurrentNetwork } from 'shared/session/selectors'
 import type { ExplorerIdentity } from 'shared/session/types'
-import { fetchOwnedENS } from 'shared/web3'
+import { fetchOwnedENS } from 'lib/web3/fetchOwnedENS'
 import { profileRequest, profileSuccess, saveProfileDelta } from '../actions'
 import { fetchProfile } from './fetchProfile'
 import { fetchLocalProfile } from './local/index'
