+++ browser-interface/packages/shared/session/sagas.ts
@@ -1,28 +1,24 @@
 import { Authenticator } from '@dcl/crypto'
-import { call, put, select, take, takeEvery, takeLatest } from 'redux-saga/effects'
-
-import { DEBUG_KERNEL_LOG, ETHEREUM_NETWORK, PREVIEW } from 'config'
-
-import { createDummyLogger, createLogger } from 'lib/logger'
-import { getUserAccount, isSessionExpired, requestManager } from 'shared/ethereum/provider'
-import { awaitingUserSignature, AWAITING_USER_SIGNATURE } from 'shared/loading/types'
-import { registerProviderNetChanges } from 'shared/web3'
-import { getEthereumNetworkFromProvider } from "shared/getEthereumNetworkFromProvider"
-
-import { getFromPersistentStorage, saveToPersistentStorage } from 'lib/browser/persistentStorage'
-
 import { createUnsafeIdentity } from '@dcl/crypto/dist/crypto'
+import { DEBUG_KERNEL_LOG, ETHEREUM_NETWORK, PREVIEW } from 'config'
 import { RequestManager } from 'eth-connect'
 import { DecentralandIdentity, LoginState } from 'kernel-web-interface'
+import { getFromPersistentStorage, saveToPersistentStorage } from 'lib/browser/persistentStorage'
+import { createDummyLogger, createLogger } from 'lib/logger'
+import { getEthereumNetworkFromProvider } from 'lib/web3/getEthereumNetworkFromProvider'
+import { getUserAccount, isSessionExpired, requestManager } from 'lib/web3/provider'
 import { Store } from 'redux'
+import { call, put, select, take, takeEvery, takeLatest } from 'redux-saga/effects'
 import { setRoomConnection } from 'shared/comms/actions'
 import { selectNetwork } from 'shared/dao/actions'
 import { getSelectedNetwork } from 'shared/dao/selectors'
+import { awaitingUserSignature, AWAITING_USER_SIGNATURE } from 'shared/loading/types'
 import { ensureMetaConfigurationInitialized } from 'shared/meta'
 import { globalObservable } from 'shared/observables'
 import { initialRemoteProfileLoad } from 'shared/profiles/sagas/initialRemoteProfileLoad'
 import { localProfilesRepo } from 'shared/profiles/sagas/local/localProfilesRepo'
 import { waitForRealm } from 'shared/realm/waitForRealmAdapter'
+import { registerProviderNetChanges } from 'shared/registerProviderNetChanges'
 import { waitForRendererInstance } from 'shared/renderer/sagas-helper'
 import { store } from 'shared/store/isolatedStore'
 import { saveProfileDelta } from '../profiles/actions'
@@ -45,7 +41,7 @@ import {
   UserAuthenticated,
   USER_AUTHENTICATED
 } from './actions'
-import { retrieveLastGuestSession, retrieveLastSessionByAddress, deleteSession, storeSession } from './index'
+import { deleteSession, retrieveLastGuestSession, retrieveLastSessionByAddress, storeSession } from './index'
 import { getCurrentIdentity, isGuestLogin } from './selectors'
 import { ExplorerIdentity, RootSessionState, SessionState, StoredSession } from './types'
