+++ browser-interface/packages/shared/tsconfig.json
@@ -7,6 +7,6 @@
     "strict": true,
     "types": ["@dcl/posix"]
   },
-  "include": ["*.ts"],
+  "include": ["*.ts", "web3/getEthereumNetworkFromProvider.ts", "web3/fetchCatalystNodesFromDAO.ts", "web3/queryGraph.ts"],
   "extends": "../../tsconfig.json"
 }
