+++ browser-interface/packages/unity-interface/BrowserInterface.ts
@@ -2,15 +2,14 @@ import { EcsMathReadOnlyQuaternion, EcsMathReadOnlyVector3 } from '@dcl/ecs-math
 
 import { Authenticator } from '@dcl/crypto'
 import { Avatar, generateLazyValidator, JSONSchema } from '@dcl/schemas'
-import {
-  DEBUG,
-  ethereumConfigurations,
-  playerHeight,
-  WORLD_EXPLORER
-} from 'config'
+import { DEBUG, ethereumConfigurations, playerHeight, WORLD_EXPLORER } from 'config'
 import { isAddress } from 'eth-connect'
 import future, { IFuture } from 'fp-future'
 import { getAuthHeaders } from 'lib/decentraland/authentication/signedFetch'
+import { arrayCleanup } from 'lib/javascript/arrayCleanup'
+import { now } from 'lib/javascript/now'
+import { defaultLogger } from 'lib/logger'
+import { fetchENSOwner } from 'lib/web3/fetchENSOwner'
 import { trackEvent } from 'shared/analytics/trackEvent'
 import { setDecentralandTime } from 'shared/apis/host/EnvironmentAPI'
 import { reportScenesAroundParcel, setHomeScene } from 'shared/atlas/actions'
@@ -21,7 +20,7 @@ import { sendMessage } from 'shared/chat/actions'
 import { sendPublicChatMessage } from 'shared/comms'
 import { changeRealm } from 'shared/dao'
 import { getSelectedNetwork } from 'shared/dao/selectors'
-import { getERC20Balance } from 'shared/ethereum/EthereumService'
+import { getERC20Balance } from 'lib/web3/EthereumService'
 import { leaveChannel, updateUserData } from 'shared/friends/actions'
 import { ensureFriendProfile } from 'shared/friends/ensureFriendProfile'
 import {
@@ -49,7 +48,6 @@ import { ReportFatalErrorWithUnityPayloadAsync } from 'shared/loading/ReportFata
 import { getLastUpdateTime } from 'shared/loading/selectors'
 import { AVATAR_LOADING_ERROR } from 'shared/loading/types'
 import { renderingActivated, renderingDectivated } from 'shared/loadingScreen/types'
-import { defaultLogger } from 'lib/logger'
 import { globalObservable } from 'shared/observables'
 import { denyPortableExperiences, removeScenePortableExperience } from 'shared/portableExperiences/actions'
 import { saveProfileDelta, sendProfileToRenderer } from 'shared/profiles/actions'
@@ -98,7 +96,6 @@ import {
   setVoiceChatVolume
 } from 'shared/voiceChat/actions'
 import { requestMediaDevice } from 'shared/voiceChat/sagas'
-import { fetchENSOwner } from 'shared/web3'
 import { rendererSignalSceneReady } from 'shared/world/actions'
 import {
   allScenesEvent,
@@ -113,8 +110,6 @@ import { setDelightedSurveyEnabled } from './delightedSurvey'
 import { fetchENSOwnerProfile } from './fetchENSOwnerProfile'
 import { GIFProcessor } from './gif-processor'
 import { getUnityInstance } from './IUnityInterface'
-import { arrayCleanup } from 'lib/javascript/arrayCleanup'
-import { now } from 'lib/javascript/now'
 
 declare const globalThis: { gifProcessor?: GIFProcessor; __debug_wearables: any }
 export const futures: Record<string, IFuture<any>> = {}
